Fix several ICF register dependency issues on Power
As part of work being done to reduce the likelihood of errors being
introduced by way of incorrect register dependencies in internal control
flow, some new asserts are being added to the Power codegen in OMR.
These new asserts have flagged several issues in OpenJ9 code regarding
our use of ICF in several places. Most of these have now been corrected.

Unfortunately, a couple of the discovered issues are turning out to be
more problematic to fix than expected. In order to allow the asserts to
be enabled ASAP to prevent introduction of new register depedency
issues, these known failures have been temporarily silenced by unmarking
the relevant code as ICF. This does not introduce any new issues, it
only serves to silence the assert on issues that are already known and
are being investigated.

Signed-off-by: Ben Thomas <ben@benthomas.ca>