Prepare Power UnresolvedDataSnippet for OMR changes
Previously, Power's UnresolvedDataSnippet took the address of the
instruction it should patch within the binary-encoded buffer. However,
the binary encoder refactors currently happening in OMR will make
unresolved-related expansion occur before binary encoding, making it
impossible to provide this information at that time. Instead, the
TR::Instruction object of the corresponding instruction will need to be
used instead.

To avoid breaking OpenJ9, it's necessary to make sure it can handle both
methods of passing the instruction until the refactor is complete. To
accomplish this, Power's UnresolvedDataSnippet will determine which
method to use at runtime for the time being by overriding
getAddressOfDataReference. Once the refactors are complete, this hack
can be removed and the TR::Instruction method can be used
unconditionally.

Signed-off-by: Ben Thomas <ben@benthomas.ca>