Use fixedLoadLabelAddressIntoReg for PPCInterfaceCallSnippet
Previously, the address of the PPCInterfaceCallSnippet data section
would be loaded in mainline code by having the snippet manually either
patch instructions in mainline code or fill in the pTOC with the correct
address.

In order to simplify this code and allow exploitation of P10 PC-relative
paddi, this code has been modified to use the fixedLoadLabelAddressIntoReg
helper to accomplish this task.

Signed-off-by: Ben Thomas <ben@benthomas.ca>