Remove code guarded by INLINE_LASTITABLE_CHECK
Previously, the Power codegen had some code for performing a check of
lastITable inline. However, this code was never enabled and a comment
noted that it caused performance issues. Since this code has been
inactive for so long, it no longer makes sense to keep it around, so it
has been removed.

Signed-off-by: Ben Thomas <ben@benthomas.ca>