Use snippet address from r12 in ppcCodePatching
Previously, the ppcCodePatching helper routine would attempt to figure
out the address of the interface call snippet's data area by examining
the instructions. However, this is a very brittle approach, since even a
minor change to these instructions - which could potentially be
introduced by a future code change or new peephole optimization - would
render this code incorrect.

A better option is to get the address that's already loaded in r12
instead. This register is not clobbered between the time that the
snippet address is initially loaded for interface dispatch and when the
patching helper would be called from Recompilation.spp, so this address
is still available at that time. By using r12 instead, a lot of
unnecessary complexity goes away and the code becomes much less brittle.

Signed-off-by: Ben Thomas <ben@benthomas.ca>